+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.1+4]
+
+* Renames arguments to variables
+
 ## [0.0.1+3]
 
 * Adds InterpolationOption